# openapi.yaml
openapi: "3.0.3"
info:
  title: Tano Blog API
  description: 个人博客后端 API 接口
  version: "1.0.0"
  contact:
    name: Tano
    url: https://tano.asia
servers:
  - url: /api/v1
    description: API v1 基础路径
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 Authorization 头部使用 Bearer 令牌
    CookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: 通过 jwt Cookie 传递令牌
  schemas:
...
    Pagination:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
          description: 总数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页数量
...
    PostReactions:
      type: object
      additionalProperties:
        type: integer
      description: "表情符号到数量的映射，如 { \"👍\": 5 }"
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        sort_order:
          type: integer
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    Series:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        cover_image:
          type: string
        sort_order:
          type: integer
    PostItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        cover_image:
          type: string
        view_count:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        author_name:
          type: string
        author:
          type: object
          properties:
            username:
              type: string
            display_name:
              type: string
            avatar_url:
              type: string
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        series:
          type: array
          items:
            $ref: '#/components/schemas/Series'
...
paths:
...
  /posts:
    get:
      tags: [文章]
      summary: 获取已发布的文章列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: category
          in: query
          schema:
            type: string
          description: 按分类 slug 筛选
        - name: tag
          in: query
          schema:
            type: string
          description: 按标签 slug 筛选
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词
      responses:
        "200":
          description: 文章列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PostItem'
...
  /posts/{slug}:
    get:
      tags: [文章]
      summary: 获取文章详情
      description: 已登录用户可查看草稿，匿名用户仅能看到已发布文章。返回内容包含文章反应统计。
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文章详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object
                  reactions:
                    $ref: '#/components/schemas/PostReactions'
                  user_emojis:
                    type: array
                    items:
                      type: string
        "404":
          description: 文章不存在
...
  /admin/posts:
    get:
      tags: [管理后台/文章]
      summary: 文章管理列表（含草稿）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published]
          description: 按状态筛选
      responses:
        "200":
          description: 文章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pagination'

    post:
      tags: [管理后台/文章]
      summary: 创建文章
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, slug]
              properties:
                title:
                  type: string
                slug:
                  type: string
                content:
                  type: string
                excerpt:
                  type: string
                cover_image:
                  type: string
                status:
                  type: string
                  enum: [draft, published]
                  default: draft
                is_top:
                  type: boolean
                  default: false
                allow_comment:
                  type: boolean
                  default: true
                author_name:
                  type: string
                category_id:
                  type: string
                  format: uuid
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                scheduled_at:
                  type: string
                  format: date-time
                  description: "定时发布时间（如 \"2026-07-01T09:00\"）"
                series_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: 文章已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object
...
  /admin/posts/{id}:
    get:
      tags: [管理后台/文章]
      summary: 获取单篇文章详情
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 文章详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object

    put:
      tags: [管理后台/文章]
      summary: 更新文章
      description: 修改前自动保存当前版本到修订历史。
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                slug:
                  type: string
                content:
                  type: string
                excerpt:
                  type: string
                cover_image:
                  type: string
                status:
                  type: string
                  enum: [draft, published]
                is_top:
                  type: boolean
                allow_comment:
                  type: boolean
                author_name:
                  type: string
                category_id:
                  type: string
                  format: uuid
                  nullable: true
                tag_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                scheduled_at:
                  type: string
                  format: date-time
                  nullable: true
                series_id:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        "200":
          description: 更新后的文章
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    type: object

    delete:
      tags: [管理后台/文章]
      summary: 删除文章
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 已删除
...
  /admin/series/{id}/posts:
    get:
      tags: [管理后台/系列]
      summary: 获取系列中的文章列表
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 文章列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostItem'

    put:
      tags: [管理后台/系列]
      summary: 批量设置系列文章
      description: 替换系列中的全部现有文章。
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [post_ids]
              properties:
                post_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: 已更新

  # ========================================
  # Admin - Site Config
  # ========================================
...